requirements.txt
# Visual Product Search Engine — Dependencies
# Install with: pip install -r requirements.txt

# Deep learning core
torch>=2.1.0
torchvision>=0.16.0

# CLIP
open-clip-torch>=2.24.0

# BLIP-2 / transformers
transformers>=4.37.0
accelerate>=0.26.0

# YOLO
ultralytics>=8.0.0

# Vector index
faiss-gpu>=1.7.4         # use faiss-cpu if no GPU: pip install faiss-cpu
# OR: pip install faiss-cpu

# Evaluation & data
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
scikit-learn>=1.3.0

# Web app
streamlit>=1.30.0

# Utilities
pandas>=2.0.0
matplotlib>=3.7.0
